requirements.txt
pip==24.0
pygame-ce==2.4.1
pytmx==3.32
setuptools==69.0.3
wheel==0.42.0
